Rewrite animated as class, spinner and add support for context managers
AnimatedDecorator
---------------
Well, this change a lot of things in code design. By the way, now is
more clear whose this can do and the behavior passing args as functions,
messages or other things. I setup correctly also the logging module
for easy debugging during the development.

All that examples are valid now:

    with animated('contextmanager'):
        time.sleep(2)

    @animated('with args')
    def with_args(*args):
        time.sleep(2)
        return args

    @animated()
    def decorated_with_called_decorator(*args):
        time.sleep(2)
        return args

    @animated
    def without_args(*args):
        time.sleep(2)
        return args

I have some trouble with multiple context manager nested and
when some context manager prints something. This doesn't works fine.
I need open issue #2 later about this.

AnimationStream
---------------

The other thing was the implementation of AnimationStream whose
can do my life more simple on the controlling for write & erase messages.
Before flush, I need fill with spaces and later use backspaces. Is
a lot of pain do this repeatly. So with animationstream I need just
do this:

	animation_stream = AnimationStream(sys.stdout)
	animation_stream.write(message)
	animation_stream.erase(message)

And always will occurs be fine, as expected.

SpinnerController, space_wave and _spinner
------------------------------------------

I rename too the StopSpinner to SpinnerController, whose control the
last_thread running, if is running, if is done and the postion & message
for handling the animation. The postion is necessary because if some
animation stop and other continues, the padding function
needs be start in the same bias.

By the way, I split the padding function as space_wave(x, a). Whose
x is your variable and a is the bias (SpinnerController.position)

In general I think is it. Besides, I added a handler for SIGINT
when the user try ctrl+c, this will call the staticmethod function
stop_animation!