Bleeding edge fixes: remove requirements-dev & animated as class
Wrapping animated as function is useless because sometimes
the user would be wants access the the public variables of controller,
whose has the last_thread running and other stuff.

requirements-dev is already put in setup.py only targeting the pypandoc
is useless. If I need, I put again this properly in soon.

By the way, the build was broken because this.

Blame me more.