Hacky code: add support for nested context managers
The implementation works fine and for now I don't will
write documentation about the black magic behind this.
Maybe I can you suffering with this because this decision.
But okay. That is the possible snippet:

    with animated('level1'):
        with animated('level2'):
            with animated('level3'):
                time.sleep(1)
            time.sleep(1)
        time.sleep(3)

I wrote with the namespaces for the contexts in the animation
whose will are joined with ' - ' char. That way the animation will be
executed in that order:

level3 - level2 - level1
level3 - level2
level3

Cool, huh? Take a look here: https://i.imgur.com/EeVnDyy.gif