Disable name-tests-test because I have some troubles with this
I was writing the tests functions and I was suffering, unfortunelly,
with the `coverage` command, whose doesn't works passing a wildcard for
me. I need so execute a script for tests, so my idea was put this on
tests/__main__.py, but this is not a *test.py file. Oh boy, fuck up
your mind!