Add more args to pre-commit and refactoring animated
Some idiomatic changes of the animated.py file
We need more changes. The code has 7.7/10 score.

The problem is the arg function. As I put as str,
in some moment, I check if the arg passed has the __name__
but str don't has. This works in some way because I check
if is callable() before set as self.func.

But this have a problem with coding design. I need check
later.