Add new methods on Animation stream
I added new methods (more splitted) on the object constructor
Animation. I did this mainly because is more easy to understand.
The other reason is because the fill_with_spaces only need count
the printable chars, so I need remove the actual string the
ANSI escape for colors.

Now we have minor problems with more bigger bars. Before this
some smalls animations bars are breaking in newlines because the
number of whitespace fills was a lot of big.